+++ scripts/v120_tree_editable.py
@@ -1,4 +1,9 @@
 from compas_view2.app import App
+from compas_view2.values import IntValue
+from compas_view2.values import FloatValue
+from compas_view2.values import StrValue
+from compas_view2.values import ListValue
+from compas_view2.values import DictValue
 
 viewer = App()
 
@@ -14,9 +19,13 @@ def on_item_edited(form, entry, column, value):
 
 
 data = [
-    {"column1": "a", "column2": {"value": 2, "options": [1, 2, 3]}, "on_item_edited": on_item_edited, "some_binded_obj": {}},
-    {"column1": "b", "column2": 2, "on_item_edited": on_item_edited},
-    {"column1": "c", "column2": 3, "on_item_edited": on_item_edited},
+    {"column1": "int1", "column2": IntValue(1), "on_item_edited": on_item_edited, "some_binded_obj": {}},
+    {"column1": "int2", "column2": IntValue(2, min=0, max=10), "on_item_edited": on_item_edited},
+    {"column1": "int3", "column2": IntValue(3, options=[0, 1, 2, 3]), "on_item_edited": on_item_edited},
+    {"column1": "float", "column2": FloatValue(0.005), "on_item_edited": on_item_edited},
+    {"column1": "str", "column2": StrValue("some texts"), "on_item_edited": on_item_edited},
+    {"column1": "list", "column2": ListValue([1, 2, 3], int), "on_item_edited": on_item_edited},
+    {"column1": "dict", "column2": DictValue({"a": 1}, int), "on_item_edited": on_item_edited},
 ]
 
 columns = [