+++ src/compas_view2/values/__init__.py
@@ -1,0 +1,8 @@
+from .value import Value  # noqa: F401
+from .boolvalue import BoolValue  # noqa: F401
+from .intvalue import IntValue  # noqa: F401
+from .floatvalue import FloatValue  # noqa: F401
+from .strvalue import StrValue  # noqa: F401
+from .listvalue import ListValue  # noqa: F401
+from .dictvalue import DictValue  # noqa: F401
+from .settings import Settings  # noqa: F401